// apps/desktop/src/ipc/generated/ProductionCastDirectory.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ProductionCastCharacterDto } from "./ProductionCastCharacterDto";
import type { ProductionCastMemberDto } from "./ProductionCastMemberDto";

export type ProductionCastDirectory = { members: Array<ProductionCastMemberDto>, characters: Array<ProductionCastCharacterDto>, };