// apps/desktop/src/ipc/generated/ProductionCastAssignArgs.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ProductionCastAssignArgs = { id: string, characterId?: string | null, characterName?: string | null, 
/**
 * Primary performer (others for the character become alternates) or an alternate.
 */
primary: boolean, };