// apps/desktop/src/ipc/generated/ProductionCrewUpdateArgs.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ProductionCrewUpdateArgs = { id: string, personName?: string | null, role?: string | null, department?: string | null, contact?: string | null, notes?: string | null, expectedRev?: number, };